requirements.txt
# Core ML
torch==2.10.0
torchvision==0.25.0
Pillow>=9.0,<12.0
ultralytics

# Data & Math
numpy==2.4.2
pandas==3.0.1
scipy==1.17.1
scikit-learn==1.8.0

# Computer Vision
opencv-python==4.13.0.92

# Plotting
matplotlib==3.10.8

# Utilities
tqdm==4.67.3
PyYAML==6.0.3
requests==2.32.5